refactor(group-buy): split service classes by responsibility to follow SRP (#99)
