+++ src/main/java/com/moogsan/moongsan_backend/domain/groupbuy/controller/GroupBuyQueryController.java
@@ -9,16 +9,12 @@ import com.moogsan.moongsan_backend.domain.groupbuy.dto.query.response.groupBuyL
 import com.moogsan.moongsan_backend.domain.groupbuy.dto.query.response.groupBuyList.ParticipatedList.ParticipatedListResponse;
 import com.moogsan.moongsan_backend.domain.groupbuy.dto.query.response.groupBuyList.WishList.WishListResponse;
 import com.moogsan.moongsan_backend.domain.groupbuy.dto.query.response.groupBuyUpdate.GroupBuyForUpdateResponse;
-import com.moogsan.moongsan_backend.domain.groupbuy.service.GroupBuyQueryService;
+import com.moogsan.moongsan_backend.domain.groupbuy.service.GroupBuyQueryService.*;
 import com.moogsan.moongsan_backend.domain.user.entity.CustomUserDetails;
-import com.moogsan.moongsan_backend.domain.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
@@ -28,12 +24,19 @@ import java.util.Optional;
 @RequiredArgsConstructor
 @RequestMapping("/api/group-buys")
 public class GroupBuyQueryController {
-    private final GroupBuyQueryService groupBuyService;
+    private final GetGroupBuyEditInfo getGroupBuyEditInfo;
+    private final GetGroupBuyDetailInfo getGroupBuyDetailInfo;
+    private final GetGroupBuyListByCursor getGroupBuyListByCursor;
+    private final GetGroupBuyWishList getGroupBuyWishList;
+    private final GetGroupBuyHostedList getGroupBuyHostedList;
+    private final GetGroupBuyParticipatedList getGroupBuyParticipatedList;
+    private final GetGroupBuyParticipantsInfo getGroupBuyParticipantsInfo;
+
 
     /// 공구 게시글 수정 전 정보 조회 SUCCESS
     @GetMapping("/{postId}/edit")
     public ResponseEntity<WrapperResponse<GroupBuyForUpdateResponse>> getGroupBuyEditInfo(@PathVariable Long postId) {
-        GroupBuyForUpdateResponse groupBuyForUpdate = groupBuyService.getGroupBuyEditInfo(postId);
+        GroupBuyForUpdateResponse groupBuyForUpdate = getGroupBuyEditInfo.getGroupBuyEditInfo(postId);
         return ResponseEntity.ok(
                 WrapperResponse.<GroupBuyForUpdateResponse>builder()
                         .message("공구 게시글 수정을 성공적으로 조회했습니다.")
@@ -50,7 +53,7 @@ public class GroupBuyQueryController {
 
         Long userId = (principal != null) ? principal.getUser().getId() : null;
 
-        DetailResponse detail = groupBuyService.getGroupBuyDetailInfo(userId, postId);
+        DetailResponse detail = getGroupBuyDetailInfo.getGroupBuyDetailInfo(userId, postId);
         return ResponseEntity.ok(
                 WrapperResponse.<DetailResponse>builder()
                         .message("공구 게시글 상세 정보를 성공적으로 조회했습니다.")
@@ -81,7 +84,7 @@ public class GroupBuyQueryController {
         }
 
         PagedResponse<BasicListResponse> pagedResponse =
-                groupBuyService.getGroupBuyListByCursor(userId, categoryId, orderBy,
+                getGroupBuyListByCursor.getGroupBuyListByCursor(userId, categoryId, orderBy,
                         cursorId, cursorCreatedAt, cursorPrice, limit);
         return ResponseEntity.ok(
                 WrapperResponse.<PagedResponse<BasicListResponse>>builder()
@@ -102,7 +105,7 @@ public class GroupBuyQueryController {
             @RequestParam(value = "cursorId", required = false) Long cursorId,
             @RequestParam(value = "limit", defaultValue = "10") Integer limit
     ) {
-        PagedResponse<WishListResponse> pagedResponse = groupBuyService.getGroupBuyWishList(
+        PagedResponse<WishListResponse> pagedResponse = getGroupBuyWishList.getGroupBuyWishList(
                 userDetails.getUser().getId(), sort, cursorCreatedAt, cursorId, limit);
         return ResponseEntity.ok(
                 WrapperResponse.<PagedResponse<WishListResponse>>builder()
@@ -120,7 +123,7 @@ public class GroupBuyQueryController {
             @RequestParam(value = "cursor", required = false) Long cursor,
             @RequestParam(value = "limit", defaultValue = "10") Integer limit
     ) {
-        PagedResponse<HostedListResponse> pagedResponse = groupBuyService.getGroupBuyHostedList(
+        PagedResponse<HostedListResponse> pagedResponse = getGroupBuyHostedList.getGroupBuyHostedList(
                 userDetails.getUser().getId(), sort, cursor, limit);
         return ResponseEntity.ok(
                 WrapperResponse.<PagedResponse<HostedListResponse>>builder()
@@ -140,7 +143,7 @@ public class GroupBuyQueryController {
             @RequestParam(value = "cursor", required = false) Long cursor,
             @RequestParam(value = "limit", defaultValue = "2147483647") Integer limit // defaultValue = "10" 추가 필요
     ) {
-        PagedResponse<ParticipatedListResponse> pagedResponse = groupBuyService.getGroupBuyParticipatedList(
+        PagedResponse<ParticipatedListResponse> pagedResponse = getGroupBuyParticipatedList.getGroupBuyParticipatedList(
                 userDetails.getUser().getId(), sort, cursorCreatedAt, cursor, limit);
         return ResponseEntity.ok(
                 WrapperResponse.<PagedResponse<ParticipatedListResponse>>builder()
@@ -155,7 +158,7 @@ public class GroupBuyQueryController {
     public ResponseEntity<WrapperResponse<ParticipantListResponse>> getGroupBuyParticipantsInfo(
             @AuthenticationPrincipal CustomUserDetails userDetails,
             @PathVariable Long postId) {
-        ParticipantListResponse participantList = groupBuyService.getGroupBuyParticipantsInfo(
+        ParticipantListResponse participantList = getGroupBuyParticipantsInfo.getGroupBuyParticipantsInfo(
                 userDetails.getUser().getId(), postId);
         return ResponseEntity.ok(
                 WrapperResponse.<ParticipantListResponse>builder()
